requirements.txt
requests
pandas
python-dotenv
beautifulsoup4
tweepy
transformers
torch
lxml
selenium
webdriver-manager
anthropic
google-generativeai 